chore: organize and tidy files
